# docs/openapi.yaml
openapi: 3.1.0
info:
  title: Anjuman e Saifee Coordination API
  version: 1.0.0
  description: API-first contract for the WhatsApp coordination portal, agent tools, admin dashboard, transcript parser, and task workflows.
servers:
  - url: /api
security:
  - whatsappPhone: []
  - portalSession: []
  - adminKey: []
...
paths:
...
  /admin/users/{id}/welcome:
    post:
      tags: [users]
      summary: (Re)send the onboarding welcome to an existing user.
      description: >-
        Sends the welcome email + `committee_platform_access_created` WhatsApp template with a
        fresh password-setup link, without touching department memberships. Always (re)sends — it
        forces past the once-per-user guard, so an admin can re-issue a welcome on request. A
        department is resolved from the user's memberships when not supplied; the user must belong
        to at least one department.
      security: [{ portalSession: [] }, { adminKey: [] }]
      parameters: [{ $ref: "#/components/parameters/Id" }]
      requestBody:
        required: false
        content:
          application/json:
            schema:
              type: object
              properties:
                department_id: { type: string, format: uuid }
      responses:
        "200":
          description: Welcome attempted.
          content:
            application/json:
              schema:
                type: object
                properties:
                  welcome_notification: { $ref: "#/components/schemas/WelcomeNotification" }
        "400": { $ref: "#/components/responses/BadRequest" }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "404": { description: User not found. }
        "500": { $ref: "#/components/responses/ServerError" }
...
components:
  securitySchemes:
    whatsappPhone:
      type: apiKey
      in: header
      name: x-whatsapp-from
      description: E.164 phone number used by the WhatsApp agent and scoped user APIs.
    adminKey:
      type: apiKey
      in: header
      name: x-admin-key
      description: Server-to-server auth for agent tools and cron jobs. Never sent by browser clients.
    portalSession:
      type: apiKey
      in: cookie
      name: portal_session
      description: HMAC-SHA256-signed httpOnly session cookie set by POST /api/admin/auth. Carries { user_id, exp }. Used by all browser-facing admin routes.
    cronSecret:
      type: http
      scheme: bearer
      description: Cron bearer token derived from CRON_SECRET.
  parameters:
    Id:
      name: id
      in: path
      required: true
      schema: { type: string, format: uuid }
...
  responses:
    BadRequest:
      description: Invalid request.
      content:
        application/json:
          schema: { $ref: "#/components/schemas/Error" }
    Unauthorized:
      description: Missing or invalid authentication.
      content:
        application/json:
          schema: { $ref: "#/components/schemas/Error" }
...
    ServerError:
      description: Server error.
      content:
        application/json:
          schema: { $ref: "#/components/schemas/Error" }
  schemas:
    Error:
      type: object
      required: [error]
      properties:
        error: { type: string }
...
    WelcomeNotification:
      type: object
      description: >-
        Outcome of an onboarding welcome attempt. `already_welcomed` is true when the user had
        already been welcomed and the automatic path therefore sent nothing (the manual
        /welcome action forces a re-send and never short-circuits).
      properties:
        email: { type: string, enum: [sent, skipped, failed] }
        whatsapp: { type: string, enum: [sent, skipped, failed] }
        already_welcomed: { type: boolean }
        set_password_url: { type: string }
        errors:
          type: array
          items: { type: string }
...